reorganize rust order events module
