--- nautilus_core/model/src/events/order.rs
@@ -1,554 +1,0 @@
-// -------------------------------------------------------------------------------------------------
-//  Copyright (C) 2015-2023 Nautech Systems Pty Ltd. All rights reserved.
-//  https://nautechsystems.io
-//
-//  Licensed under the GNU Lesser General Public License Version 3.0 (the "License");
-//  You may not use this file except in compliance with the License.
-//  You may obtain a copy of the License at https://www.gnu.org/licenses/lgpl-3.0.en.html
-//
-//  Unless required by applicable law or agreed to in writing, software
-//  distributed under the License is distributed on an "AS IS" BASIS,
-//  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-//  See the License for the specific language governing permissions and
-//  limitations under the License.
-// -------------------------------------------------------------------------------------------------
-
-use std::{
-    collections::HashMap,
-    fmt::{Display, Formatter},
-};
-
-use anyhow::Result;
-use derive_builder::{self, Builder};
-use nautilus_core::{time::UnixNanos, uuid::UUID4};
-use pyo3::prelude::*;
-use serde::{Deserialize, Serialize};
-use ustr::Ustr;
-
-use crate::{
-    enums::{
-        ContingencyType, LiquiditySide, OrderSide, OrderType, TimeInForce, TrailingOffsetType,
-        TriggerType,
-    },
-    identifiers::{
-        account_id::AccountId, client_order_id::ClientOrderId, exec_algorithm_id::ExecAlgorithmId,
-        instrument_id::InstrumentId, order_list_id::OrderListId, position_id::PositionId,
-        strategy_id::StrategyId, trade_id::TradeId, trader_id::TraderId,
-        venue_order_id::VenueOrderId,
-    },
-    types::{currency::Currency, money::Money, price::Price, quantity::Quantity},
-};
-
-#[derive(Clone, PartialEq, Eq, Debug, Serialize, Deserialize)]
-pub enum OrderEvent {
-    OrderInitialized(Box<OrderInitialized>),
-    OrderDenied(OrderDenied),
-    OrderEmulated(OrderEmulated),
-    OrderReleased(OrderReleased),
-    OrderSubmitted(OrderSubmitted),
-    OrderAccepted(OrderAccepted),
-    OrderRejected(OrderRejected),
-    OrderCanceled(OrderCanceled),
-    OrderExpired(OrderExpired),
-    OrderTriggered(OrderTriggered),
-    OrderPendingUpdate(OrderPendingUpdate),
-    OrderPendingCancel(OrderPendingCancel),
-    OrderModifyRejected(OrderModifyRejected),
-    OrderCancelRejected(OrderCancelRejected),
-    OrderUpdated(OrderUpdated),
-    OrderPartiallyFilled(OrderFilled),
-    OrderFilled(OrderFilled),
-}
-
-impl OrderEvent {
-    #[must_use]
-    pub fn client_order_id(&self) -> ClientOrderId {
-        match self {
-            Self::OrderInitialized(e) => e.client_order_id,
-            Self::OrderDenied(e) => e.client_order_id,
-            Self::OrderEmulated(e) => e.client_order_id,
-            Self::OrderReleased(e) => e.client_order_id,
-            Self::OrderSubmitted(e) => e.client_order_id,
-            Self::OrderAccepted(e) => e.client_order_id,
-            Self::OrderRejected(e) => e.client_order_id,
-            Self::OrderCanceled(e) => e.client_order_id,
-            Self::OrderExpired(e) => e.client_order_id,
-            Self::OrderTriggered(e) => e.client_order_id,
-            Self::OrderPendingUpdate(e) => e.client_order_id,
-            Self::OrderPendingCancel(e) => e.client_order_id,
-            Self::OrderModifyRejected(e) => e.client_order_id,
-            Self::OrderCancelRejected(e) => e.client_order_id,
-            Self::OrderUpdated(e) => e.client_order_id,
-            Self::OrderPartiallyFilled(e) => e.client_order_id,
-            Self::OrderFilled(e) => e.client_order_id,
-        }
-    }
-
-    #[must_use]
-    pub fn strategy_id(&self) -> StrategyId {
-        match self {
-            Self::OrderInitialized(e) => e.strategy_id,
-            Self::OrderDenied(e) => e.strategy_id,
-            Self::OrderEmulated(e) => e.strategy_id,
-            Self::OrderReleased(e) => e.strategy_id,
-            Self::OrderSubmitted(e) => e.strategy_id,
-            Self::OrderAccepted(e) => e.strategy_id,
-            Self::OrderRejected(e) => e.strategy_id,
-            Self::OrderCanceled(e) => e.strategy_id,
-            Self::OrderExpired(e) => e.strategy_id,
-            Self::OrderTriggered(e) => e.strategy_id,
-            Self::OrderPendingUpdate(e) => e.strategy_id,
-            Self::OrderPendingCancel(e) => e.strategy_id,
-            Self::OrderModifyRejected(e) => e.strategy_id,
-            Self::OrderCancelRejected(e) => e.strategy_id,
-            Self::OrderUpdated(e) => e.strategy_id,
-            Self::OrderPartiallyFilled(e) => e.strategy_id,
-            Self::OrderFilled(e) => e.strategy_id,
-        }
-    }
-
-    #[must_use]
-    pub fn ts_event(&self) -> UnixNanos {
-        match self {
-            Self::OrderInitialized(e) => e.ts_event,
-            Self::OrderDenied(e) => e.ts_event,
-            Self::OrderEmulated(e) => e.ts_event,
-            Self::OrderReleased(e) => e.ts_event,
-            Self::OrderSubmitted(e) => e.ts_event,
-            Self::OrderAccepted(e) => e.ts_event,
-            Self::OrderRejected(e) => e.ts_event,
-            Self::OrderCanceled(e) => e.ts_event,
-            Self::OrderExpired(e) => e.ts_event,
-            Self::OrderTriggered(e) => e.ts_event,
-            Self::OrderPendingUpdate(e) => e.ts_event,
-            Self::OrderPendingCancel(e) => e.ts_event,
-            Self::OrderModifyRejected(e) => e.ts_event,
-            Self::OrderCancelRejected(e) => e.ts_event,
-            Self::OrderUpdated(e) => e.ts_event,
-            Self::OrderPartiallyFilled(e) => e.ts_event,
-            Self::OrderFilled(e) => e.ts_event,
-        }
-    }
-}
-
-#[repr(C)]
-#[derive(Clone, PartialEq, Eq, Debug, Builder, Serialize, Deserialize)]
-#[builder(default)]
-#[serde(tag = "type")]
-pub struct OrderInitialized {
-    pub trader_id: TraderId,
-    pub strategy_id: StrategyId,
-    pub instrument_id: InstrumentId,
-    pub client_order_id: ClientOrderId,
-    pub order_side: OrderSide,
-    pub order_type: OrderType,
-    pub quantity: Quantity,
-    pub price: Option<Price>,
-    pub trigger_price: Option<Price>,
-    pub trigger_type: Option<TriggerType>,
-    pub limit_offset: Option<Price>,
-    pub trailing_offset: Option<Price>,
-    pub trailing_offset_type: Option<TrailingOffsetType>,
-    pub time_in_force: TimeInForce,
-    pub expire_time: Option<UnixNanos>,
-    pub post_only: bool,
-    pub reduce_only: bool,
-    pub quote_quantity: bool,
-    pub display_qty: Option<Quantity>,
-    pub emulation_trigger: Option<TriggerType>,
-    pub trigger_instrument_id: Option<InstrumentId>,
-    pub contingency_type: Option<ContingencyType>,
-    pub order_list_id: Option<OrderListId>,
-    pub linked_order_ids: Option<Vec<ClientOrderId>>,
-    pub parent_order_id: Option<ClientOrderId>,
-    pub exec_algorithm_id: Option<ExecAlgorithmId>,
-    pub exec_algorithm_params: Option<HashMap<Ustr, Ustr>>,
-    pub exec_spawn_id: Option<ClientOrderId>,
-    pub tags: Option<Ustr>,
-    pub event_id: UUID4,
-    pub ts_event: UnixNanos,
-    pub ts_init: UnixNanos,
-    pub reconciliation: bool,
-}
-
-impl Default for OrderInitialized {
-    fn default() -> Self {
-        Self {
-            trader_id: TraderId::default(),
-            strategy_id: StrategyId::default(),
-            instrument_id: InstrumentId::default(),
-            client_order_id: ClientOrderId::default(),
-            order_side: OrderSide::Buy,
-            order_type: OrderType::Market,
-            quantity: Quantity::new(100_000.0, 0).unwrap(),
-            price: Default::default(),
-            trigger_price: Default::default(),
-            trigger_type: Default::default(),
-            time_in_force: TimeInForce::Day,
-            expire_time: Default::default(),
-            post_only: Default::default(),
-            reduce_only: Default::default(),
-            display_qty: Default::default(),
-            quote_quantity: Default::default(),
-            limit_offset: Default::default(),
-            trailing_offset: Default::default(),
-            trailing_offset_type: Default::default(),
-            emulation_trigger: Default::default(),
-            trigger_instrument_id: Default::default(),
-            contingency_type: Default::default(),
-            order_list_id: Default::default(),
-            linked_order_ids: Default::default(),
-            parent_order_id: Default::default(),
-            exec_algorithm_id: Default::default(),
-            exec_algorithm_params: Default::default(),
-            exec_spawn_id: Default::default(),
-            tags: Default::default(),
-            event_id: Default::default(),
-            ts_event: Default::default(),
-            ts_init: Default::default(),
-            reconciliation: Default::default(),
-        }
-    }
-}
-
-#[repr(C)]
-#[derive(Clone, Copy, PartialEq, Eq, Debug, Default, Serialize, Deserialize, Builder)]
-#[builder(default)]
-#[serde(tag = "type")]
-#[cfg_attr(
-    feature = "python",
-    pyclass(module = "nautilus_trader.core.nautilus_pyo3.model")
-)]
-pub struct OrderDenied {
-    pub trader_id: TraderId,
-    pub strategy_id: StrategyId,
-    pub instrument_id: InstrumentId,
-    pub client_order_id: ClientOrderId,
-    pub reason: Ustr,
-    pub event_id: UUID4,
-    pub ts_event: UnixNanos,
-    pub ts_init: UnixNanos,
-}
-
-impl OrderDenied {
-    #[allow(clippy::too_many_arguments)]
-    pub fn new(
-        trader_id: TraderId,
-        strategy_id: StrategyId,
-        instrument_id: InstrumentId,
-        client_order_id: ClientOrderId,
-        reason: Ustr,
-        event_id: UUID4,
-        ts_event: UnixNanos,
-        ts_init: UnixNanos,
-    ) -> Result<OrderDenied> {
-        Ok(OrderDenied {
-            trader_id,
-            strategy_id,
-            instrument_id,
-            client_order_id,
-            reason,
-            event_id,
-            ts_event,
-            ts_init,
-        })
-    }
-}
-
-impl Display for OrderDenied {
-    fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
-        write!(
-            f,
-            "OrderDenied(instrument_id={}, client_order_id={}, reason={})",
-            self.instrument_id, self.client_order_id, self.reason
-        )
-    }
-}
-
-#[repr(C)]
-#[derive(Clone, Copy, PartialEq, Eq, Debug, Default, Serialize, Deserialize, Builder)]
-#[builder(default)]
-#[serde(tag = "type")]
-pub struct OrderEmulated {
-    pub trader_id: TraderId,
-    pub strategy_id: StrategyId,
-    pub instrument_id: InstrumentId,
-    pub client_order_id: ClientOrderId,
-    pub event_id: UUID4,
-    pub ts_event: UnixNanos,
-    pub ts_init: UnixNanos,
-}
-
-#[repr(C)]
-#[derive(Clone, Copy, PartialEq, Eq, Debug, Default, Serialize, Deserialize, Builder)]
-#[builder(default)]
-#[serde(tag = "type")]
-pub struct OrderReleased {
-    pub trader_id: TraderId,
-    pub strategy_id: StrategyId,
-    pub instrument_id: InstrumentId,
-    pub client_order_id: ClientOrderId,
-    pub released_price: Price,
-    pub event_id: UUID4,
-    pub ts_event: UnixNanos,
-    pub ts_init: UnixNanos,
-}
-
-#[repr(C)]
-#[derive(Clone, PartialEq, Eq, Debug, Default, Serialize, Deserialize, Builder)]
-#[builder(default)]
-#[serde(tag = "type")]
-pub struct OrderSubmitted {
-    pub trader_id: TraderId,
-    pub strategy_id: StrategyId,
-    pub instrument_id: InstrumentId,
-    pub client_order_id: ClientOrderId,
-    pub account_id: AccountId,
-    pub event_id: UUID4,
-    pub ts_event: UnixNanos,
-    pub ts_init: UnixNanos,
-}
-
-#[repr(C)]
-#[derive(Clone, PartialEq, Eq, Debug, Default, Serialize, Deserialize, Builder)]
-#[builder(default)]
-#[serde(tag = "type")]
-pub struct OrderAccepted {
-    pub trader_id: TraderId,
-    pub strategy_id: StrategyId,
-    pub instrument_id: InstrumentId,
-    pub client_order_id: ClientOrderId,
-    pub venue_order_id: VenueOrderId,
-    pub account_id: AccountId,
-    pub event_id: UUID4,
-    pub ts_event: UnixNanos,
-    pub ts_init: UnixNanos,
-    pub reconciliation: u8,
-}
-
-#[repr(C)]
-#[derive(Clone, PartialEq, Eq, Debug, Default, Serialize, Deserialize, Builder)]
-#[builder(default)]
-#[serde(tag = "type")]
-pub struct OrderRejected {
-    pub trader_id: TraderId,
-    pub strategy_id: StrategyId,
-    pub instrument_id: InstrumentId,
-    pub client_order_id: ClientOrderId,
-    pub account_id: AccountId,
-    pub reason: Ustr,
-    pub event_id: UUID4,
-    pub ts_event: UnixNanos,
-    pub ts_init: UnixNanos,
-    pub reconciliation: u8,
-}
-
-#[repr(C)]
-#[derive(Clone, PartialEq, Eq, Debug, Default, Serialize, Deserialize, Builder)]
-#[builder(default)]
-#[serde(tag = "type")]
-pub struct OrderCanceled {
-    pub trader_id: TraderId,
-    pub strategy_id: StrategyId,
-    pub instrument_id: InstrumentId,
-    pub client_order_id: ClientOrderId,
-    pub venue_order_id: Option<VenueOrderId>,
-    pub account_id: Option<AccountId>,
-    pub event_id: UUID4,
-    pub ts_event: UnixNanos,
-    pub ts_init: UnixNanos,
-    pub reconciliation: bool,
-}
-
-#[repr(C)]
-#[derive(Clone, PartialEq, Eq, Debug, Default, Serialize, Deserialize, Builder)]
-#[builder(default)]
-#[serde(tag = "type")]
-pub struct OrderExpired {
-    pub trader_id: TraderId,
-    pub strategy_id: StrategyId,
-    pub instrument_id: InstrumentId,
-    pub client_order_id: ClientOrderId,
-    pub venue_order_id: Option<VenueOrderId>,
-    pub account_id: Option<AccountId>,
-    pub event_id: UUID4,
-    pub ts_event: UnixNanos,
-    pub ts_init: UnixNanos,
-    pub reconciliation: bool,
-}
-
-#[repr(C)]
-#[derive(Clone, PartialEq, Eq, Debug, Default, Serialize, Deserialize, Builder)]
-#[builder(default)]
-#[serde(tag = "type")]
-pub struct OrderTriggered {
-    pub trader_id: TraderId,
-    pub strategy_id: StrategyId,
-    pub instrument_id: InstrumentId,
-    pub client_order_id: ClientOrderId,
-    pub venue_order_id: Option<VenueOrderId>,
-    pub account_id: Option<AccountId>,
-    pub event_id: UUID4,
-    pub ts_event: UnixNanos,
-    pub ts_init: UnixNanos,
-    pub reconciliation: bool,
-}
-
-#[repr(C)]
-#[derive(Clone, PartialEq, Eq, Debug, Default, Serialize, Deserialize, Builder)]
-#[builder(default)]
-#[serde(tag = "type")]
-pub struct OrderPendingUpdate {
-    pub trader_id: TraderId,
-    pub strategy_id: StrategyId,
-    pub instrument_id: InstrumentId,
-    pub client_order_id: ClientOrderId,
-    pub venue_order_id: Option<VenueOrderId>,
-    pub account_id: AccountId,
-    pub event_id: UUID4,
-    pub ts_event: UnixNanos,
-    pub ts_init: UnixNanos,
-    pub reconciliation: bool,
-}
-
-#[repr(C)]
-#[derive(Clone, PartialEq, Eq, Debug, Default, Serialize, Deserialize, Builder)]
-#[builder(default)]
-#[serde(tag = "type")]
-pub struct OrderPendingCancel {
-    pub trader_id: TraderId,
-    pub strategy_id: StrategyId,
-    pub instrument_id: InstrumentId,
-    pub client_order_id: ClientOrderId,
-    pub venue_order_id: Option<VenueOrderId>,
-    pub account_id: AccountId,
-    pub event_id: UUID4,
-    pub ts_event: UnixNanos,
-    pub ts_init: UnixNanos,
-    pub reconciliation: bool,
-}
-
-#[repr(C)]
-#[derive(Clone, PartialEq, Eq, Debug, Default, Serialize, Deserialize, Builder)]
-#[builder(default)]
-#[serde(tag = "type")]
-pub struct OrderModifyRejected {
-    pub trader_id: TraderId,
-    pub strategy_id: StrategyId,
-    pub instrument_id: InstrumentId,
-    pub client_order_id: ClientOrderId,
-    pub venue_order_id: Option<VenueOrderId>,
-    pub account_id: Option<AccountId>,
-    pub reason: Ustr,
-    pub event_id: UUID4,
-    pub ts_event: UnixNanos,
-    pub ts_init: UnixNanos,
-    pub reconciliation: bool,
-}
-
-#[repr(C)]
-#[derive(Clone, PartialEq, Eq, Debug, Default, Serialize, Deserialize, Builder)]
-#[builder(default)]
-#[serde(tag = "type")]
-pub struct OrderCancelRejected {
-    pub trader_id: TraderId,
-    pub strategy_id: StrategyId,
-    pub instrument_id: InstrumentId,
-    pub client_order_id: ClientOrderId,
-    pub venue_order_id: Option<VenueOrderId>,
-    pub account_id: Option<AccountId>,
-    pub reason: Ustr,
-    pub event_id: UUID4,
-    pub ts_event: UnixNanos,
-    pub ts_init: UnixNanos,
-    pub reconciliation: bool,
-}
-
-#[repr(C)]
-#[derive(Clone, PartialEq, Eq, Debug, Default, Serialize, Deserialize, Builder)]
-#[builder(default)]
-#[serde(tag = "type")]
-pub struct OrderUpdated {
-    pub trader_id: TraderId,
-    pub strategy_id: StrategyId,
-    pub instrument_id: InstrumentId,
-    pub client_order_id: ClientOrderId,
-    pub venue_order_id: Option<VenueOrderId>,
-    pub account_id: Option<AccountId>,
-    pub quantity: Quantity,
-    pub price: Option<Price>,
-    pub trigger_price: Option<Price>,
-    pub event_id: UUID4,
-    pub ts_event: UnixNanos,
-    pub ts_init: UnixNanos,
-    pub reconciliation: bool,
-}
-
-#[repr(C)]
-#[derive(Clone, PartialEq, Eq, Debug, Serialize, Deserialize, Builder)]
-#[builder(default)]
-#[serde(tag = "type")]
-pub struct OrderFilled {
-    pub trader_id: TraderId,
-    pub strategy_id: StrategyId,
-    pub instrument_id: InstrumentId,
-    pub client_order_id: ClientOrderId,
-    pub venue_order_id: VenueOrderId,
-    pub account_id: AccountId,
-    pub trade_id: TradeId,
-    pub position_id: Option<PositionId>,
-    pub order_side: OrderSide,
-    pub order_type: OrderType,
-    pub last_qty: Quantity,
-    pub last_px: Price,
-    pub currency: Currency,
-    pub commission: Option<Money>,
-    pub liquidity_side: LiquiditySide,
-    pub event_id: UUID4,
-    pub ts_event: UnixNanos,
-    pub ts_init: UnixNanos,
-    pub reconciliation: bool,
-}
-
-impl Default for OrderFilled {
-    fn default() -> Self {
-        Self {
-            trader_id: TraderId::default(),
-            strategy_id: StrategyId::default(),
-            instrument_id: InstrumentId::default(),
-            client_order_id: ClientOrderId::default(),
-            venue_order_id: VenueOrderId::default(),
-            account_id: AccountId::default(),
-            trade_id: TradeId::default(),
-            position_id: None,
-            order_side: OrderSide::Buy,
-            order_type: OrderType::Market,
-            last_qty: Quantity::new(100_000.0, 0).unwrap(),
-            last_px: Price::from("1.00000"),
-            currency: Currency::USD(),
-            commission: None,
-            liquidity_side: LiquiditySide::Taker,
-            event_id: Default::default(),
-            ts_event: Default::default(),
-            ts_init: Default::default(),
-            reconciliation: Default::default(),
-        }
-    }
-}
-
-////////////////////////////////////////////////////////////////////////////////
-// Tests
-////////////////////////////////////////////////////////////////////////////////
-#[cfg(test)]
-mod tests {
-    use rstest::rstest;
-
-    use super::*;
-    use crate::events::stubs::*;
-
-    #[rstest]
-    fn test_order_denied_display(order_denied_max_submitted_rate: OrderDenied) {
-        let display = format!("{}", order_denied_max_submitted_rate);
-        assert_eq!(display, "OrderDenied(instrument_id=BTCUSDT.COINBASE, client_order_id=O-20200814-102234-001-001-1, reason=Exceeded MAX_ORDER_SUBMIT_RATE)");
-    }
-}
