+++ nautilus_core/model/src/events/order/mod.rs
@@ -13,5 +13,22 @@
 //  limitations under the License.
 // -------------------------------------------------------------------------------------------------
 
-pub mod order;
-pub mod position;
+pub mod order_accepted;
+pub mod order_cancel_rejected;
+pub mod order_canceled;
+pub mod order_denied;
+pub mod order_emulated;
+pub mod order_event;
+pub mod order_expired;
+pub mod order_filled;
+pub mod order_initialized;
+pub mod order_modified_rejected;
+pub mod order_pending_cancel;
+pub mod order_pending_update;
+pub mod order_rejected;
+pub mod order_released;
+pub mod order_submitted;
+pub mod order_triggered;
+pub mod order_updated;
+#[cfg(feature = "stubs")]
+pub mod stubs;
