+++ nautilus_core/model/src/events/order/order_filled.rs
@@ -1,0 +1,80 @@
+// -------------------------------------------------------------------------------------------------
+//  Copyright (C) 2015-2023 Nautech Systems Pty Ltd. All rights reserved.
+//  https://nautechsystems.io
+//
+//  Licensed under the GNU Lesser General Public License Version 3.0 (the "License");
+//  You may not use this file except in compliance with the License.
+//  You may obtain a copy of the License at https://www.gnu.org/licenses/lgpl-3.0.en.html
+//
+//  Unless required by applicable law or agreed to in writing, software
+//  distributed under the License is distributed on an "AS IS" BASIS,
+//  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+//  See the License for the specific language governing permissions and
+//  limitations under the License.
+// -------------------------------------------------------------------------------------------------
+
+use derive_builder::{self, Builder};
+use nautilus_core::{time::UnixNanos, uuid::UUID4};
+use serde::{Deserialize, Serialize};
+
+use crate::{
+    enums::{LiquiditySide, OrderSide, OrderType},
+    identifiers::{
+        account_id::AccountId, client_order_id::ClientOrderId, instrument_id::InstrumentId,
+        position_id::PositionId, strategy_id::StrategyId, trade_id::TradeId, trader_id::TraderId,
+        venue_order_id::VenueOrderId,
+    },
+    types::{currency::Currency, money::Money, price::Price, quantity::Quantity},
+};
+
+#[repr(C)]
+#[derive(Clone, PartialEq, Eq, Debug, Serialize, Deserialize, Builder)]
+#[builder(default)]
+#[serde(tag = "type")]
+pub struct OrderFilled {
+    pub trader_id: TraderId,
+    pub strategy_id: StrategyId,
+    pub instrument_id: InstrumentId,
+    pub client_order_id: ClientOrderId,
+    pub venue_order_id: VenueOrderId,
+    pub account_id: AccountId,
+    pub trade_id: TradeId,
+    pub position_id: Option<PositionId>,
+    pub order_side: OrderSide,
+    pub order_type: OrderType,
+    pub last_qty: Quantity,
+    pub last_px: Price,
+    pub currency: Currency,
+    pub commission: Option<Money>,
+    pub liquidity_side: LiquiditySide,
+    pub event_id: UUID4,
+    pub ts_event: UnixNanos,
+    pub ts_init: UnixNanos,
+    pub reconciliation: bool,
+}
+
+impl Default for OrderFilled {
+    fn default() -> Self {
+        Self {
+            trader_id: TraderId::default(),
+            strategy_id: StrategyId::default(),
+            instrument_id: InstrumentId::default(),
+            client_order_id: ClientOrderId::default(),
+            venue_order_id: VenueOrderId::default(),
+            account_id: AccountId::default(),
+            trade_id: TradeId::default(),
+            position_id: None,
+            order_side: OrderSide::Buy,
+            order_type: OrderType::Market,
+            last_qty: Quantity::new(100_000.0, 0).unwrap(),
+            last_px: Price::from("1.00000"),
+            currency: Currency::USD(),
+            commission: None,
+            liquidity_side: LiquiditySide::Taker,
+            event_id: Default::default(),
+            ts_event: Default::default(),
+            ts_init: Default::default(),
+            reconciliation: Default::default(),
+        }
+    }
+}
