+++ nautilus_core/model/src/events/order/order_released.rs
@@ -13,5 +13,29 @@
 //  limitations under the License.
 // -------------------------------------------------------------------------------------------------
 
-pub mod order;
-pub mod position;
+use derive_builder::{self, Builder};
+use nautilus_core::{time::UnixNanos, uuid::UUID4};
+use serde::{Deserialize, Serialize};
+
+use crate::{
+    identifiers::{
+        client_order_id::ClientOrderId, instrument_id::InstrumentId, strategy_id::StrategyId,
+        trader_id::TraderId,
+    },
+    types::price::Price,
+};
+
+#[repr(C)]
+#[derive(Clone, Copy, PartialEq, Eq, Debug, Default, Serialize, Deserialize, Builder)]
+#[builder(default)]
+#[serde(tag = "type")]
+pub struct OrderReleased {
+    pub trader_id: TraderId,
+    pub strategy_id: StrategyId,
+    pub instrument_id: InstrumentId,
+    pub client_order_id: ClientOrderId,
+    pub released_price: Price,
+    pub event_id: UUID4,
+    pub ts_event: UnixNanos,
+    pub ts_init: UnixNanos,
+}
