+++ nautilus_core/model/src/ffi/events/order.rs
@@ -19,7 +19,9 @@ use nautilus_core::{ffi::string::cstr_to_ustr, time::UnixNanos, uuid::UUID4};
 
 use crate::{
     events::order::{
-        OrderAccepted, OrderDenied, OrderEmulated, OrderRejected, OrderReleased, OrderSubmitted,
+        order_accepted::OrderAccepted, order_denied::OrderDenied, order_emulated::OrderEmulated,
+        order_rejected::OrderRejected, order_released::OrderReleased,
+        order_submitted::OrderSubmitted,
     },
     identifiers::{
         account_id::AccountId, client_order_id::ClientOrderId, instrument_id::InstrumentId,
